% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_with_pause.R
\name{gene_summary_with_pause}
\alias{gene_summary_with_pause}
\title{Title}
\usage{
gene_summary_with_pause(gene, timesec = 2, ...)
}
\arguments{
\item{gene}{和gene_summary的参数一样，也是entrezID}

\item{timesec}{停留时间为秒，默认2秒}

\item{...}{跟随fanyi包函数gene_summary中的其他可能参数参数}
}
\description{
Title
}
