% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pkgs.R
\name{install_packages_with_pak}
\alias{install_packages_with_pak}
\title{Title}
\usage{
install_packages_with_pak(packages)
}
\arguments{
\item{packages}{this is a must packages c("pak","fanyi","clusterProfiler","enrichplot")}
}
\description{
Title
}
