% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_offical}
\alias{kegg_offical}
\title{kegg_offical}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54789 rows and 9 columns.
}
\source{
Check KEGG Offical website inhere \href{https://www.genome.jp/kegg-bin/show_organism?menu_type=gene_catalogs&org=hsa}{hsa@KEGG}
}
\usage{
kegg_offical
}
\description{
The `kegg_offical` dataset comes from the KEGG official website, which includes all human pathways and pathway level classifications from the official website, such as L1 for level 1 pathways, L2 for level 2 pathways, and L3 for level 3 pathways.
}
\details{
This is not available in other databases.
The original data processing process can be found in the `data_gson.R` file.
}
\examples{
data(kegg_offical)

}
\keyword{datasets}
