% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myenricher.R
\name{myenricher}
\alias{myenricher}
\title{Title}
\usage{
myenricher(def = NULL, gson = NULL, gene = genes, ...)
}
\arguments{
\item{def}{The "def" is to be consistent with the results obtained by the company, as they filtered the meridians based on the selected genes.
Perhaps my understanding is incorrect, but Y's results also came out after filtering, but the p-value is significant. After filtering, the p-value is no longer significant.
The default parameter is "luckcp".}

\item{gson}{"gson" is your multiple lists, of course, one of them may have already been processed.}

\item{gene}{Your list of significant differentially expressed genes, or your list of genes that you consider important.}

\item{...}{Other parameters can be acctpted by the “enricher” function}
}
\description{
Title
}
