% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_with_pause.R
\name{translate_with_pause}
\alias{translate_with_pause}
\title{Title}
\usage{
translate_with_pause(timesec = 2, ...)
}
\arguments{
\item{timesec}{停留时间为秒，默认2秒}

\item{...}{跟随fanyi包函数translate中的其他可能参数参数}
}
\value{
No return value, called for side effects
}
\description{
Title
}
